% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_diff.R
\name{calc_diff}
\alias{calc_diff}
\title{Calculate difference from best}
\usage{
calc_diff(x)
}
\arguments{
\item{d}{}
}
\value{
columns with raced rates that have differences from the best rate
}
\description{
Calculate difference from best
}
\examples{
calc_diff(d)

}
